рефакторинг

+++ internal/models/user.go
@@ -8,6 +8,6 @@ type User struct {
 	gorm.Model
 	Login     string  `gorm:"size:255;not null;unique" json:"login"`
 	Password  string  `gorm:"size:255;not null;unique" json:"password"`
-	Balance   float32 `json:"-"`
-	Withdrawn float32 `json:"-"`
+	Balance   float64 `json:"-"`
+	Withdrawn float64 `json:"-"`
 }
